requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
joblib>=1.3.0
pytest>=8.0.0
